fix: remove panic=abort, fix unnecessary cast lint
- panic=abort causes html2md crate to fail (needs panic_unwind)
- remove redundant  casts on already-i32 values
- CI clippy now passes with -D warnings

Constraint: CI uses strict linting
Confidence: high
Scope-risk: narrow

+++ crates/session/src/db.rs
@@ -166,7 +166,7 @@ impl SessionDB {
                     session_id,
                     event.event_type,
                     category_to_string(&event.category),
-                    event.priority as i32,
+                    event.priority,
                     event.data,
                     project_dir,
                     attribution_source,
